pyproject.toml
[tool.poetry]
name = "excel-indexer"
version = "0.1.0"
description = ""
authors = ["nuin <nuin@genedrift.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.1.7"
rich = "^13.8.1"
whoosh = "^2.7.4"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
pytesseract = "^0.3.13"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
